before last commit

+++ app/Http/Controllers/ManagerController.php
@@ -73,6 +73,16 @@ class ManagerController extends Controller
     public function showRoom()
     {
         $rooms = Room::all();
+        // $floorName = new Floor();
+        // $floorName = Floor::select($floorName->name)->where(
+        //     $room->floor_number = $floorName->id
+        // );
+        $rooms = Room::with('RelatedFloor')->get();
+        // foreach ($rooms as $r) {
+        // $floorName = $rooms->RelatedFloor->name;
+        // }
+        // $product->category->name; // get category name
+        // $product->category->user->name;
         return view('manager.show_rooms', [
             'rooms' => $rooms,
         ]);
@@ -80,9 +90,9 @@ class ManagerController extends Controller
     public function deleteRoom($number)
     {
         $room = Room::find($number);
-        // if ($room->status == 'free') {
-        Room::where('number', $number)->delete();
-        // }
+        if ($room->status == 'free') {
+            Room::where('number', $number)->delete();
+        }
         return redirect()->route('show_rooms');
     }
     public function updateRoom($RoomId)
